--- src/shared/Title/type.ts
@@ -1,11 +1,0 @@
-import React from 'react';
-
-export interface TitleInterface {
-  title: string;
-  align?: 'start' | 'center';
-  size?: number;
-  subtitle?: string;
-  subSize?: number;
-  gap?: number;
-  children: React.ReactNode;
-}
